enhance, cleanup, rename chain agnostic permission utilities
